backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
pyyaml==6.0.1
httpx==0.25.2
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
cryptography==42.0.5
python-telegram-bot==20.7
asyncio-mqtt==0.16.1
psutil==5.9.6
schedule==1.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
